WPOS-77 - mod_coursecertificate issuing scheduled task throws exception on already created file
When processing certificates for a site with many users, if a user
generated their certificate manually during the process, a file
already exists error would be generated when processing the task.

A check has been included to see if the file exists before creating
it, if it exists we return it.